Actualizar interceptor de autenticación para incluir token de acceso

+++ src/app/data/Interceptors/auth.interceptor.ts
@@ -6,23 +6,28 @@ import {app} from "../../config/App";
 
 export const authInterceptor: HttpInterceptorFn = (req, next) => {
 	const router = inject(Router);
+	let response = next(req);
 
 	const access_token: string | null = localStorage.getItem('access_token');
 
-	const authReq = req.clone({
-		setHeaders: {
-			Authorization: `Bearer ${access_token}`
-		}
-	});
-
-	return next(authReq).pipe(
-		catchError((error) => {
-			if (error.status === 401) {
-				localStorage.clear();
-				router.navigate([app.redirectLogout]);
+	if(access_token) {
+		const authReq = req.clone({
+			setHeaders: {
+				Authorization: `Bearer ${access_token}`
 			}
+		});
 
-			throw error;
-		})
-	);
+		response = next(authReq).pipe(
+			catchError((error) => {
+				if (error.status === 401) {
+					localStorage.clear();
+					router.navigate([app.redirectLogout]);
+				}
+
+				throw error;
+			})
+		);
+	}
+
+	return response;
 };
